feat: change concept for AUR lookups
Instead of fetching package info data when a package is highted,
the lookup is now being performed immediately after searching.
Once the search result is returned, another request is being made to
retrieve data for ALL packages returned by the search

Technically this seems to be the better approach than firing off several
search queries for individual packages.

See issue #3 for more information on this topic.