feat: improve execution of "commands"
The way commands are run has been improved by running the command
like $SHELL -c "whatever the user has entered".
This allows more sophisticated commands to be entered.

Next to that, instead of prepending the entered command,
you can use {pkg} within the command text.
This term will be replaced with the package name when executed.
(This is valid for Install and Uninstall commands)