feat: show list of installed packages
CTRL+L or entering en empty search string will show all installed packages.

Next to that pacseek can be started immediately displaying
all installed packages via "-i" flag.

The -i flag was previously used for the ASCII mode.
This has been changed to "-a" now.
Hope this does not create to many problems...

Next to the above, an option has been added to turn "Required by" on or off.
That's because the computation for that information is quite heavy.
The default value is "off".

Signed-off-by: moson-mo <mo-son@mailbox.org>